Refactors KaufmanEfficiencyRatio and KaufmanAdaptiveMovingAverage
